+++ Tasks/13Oct2023/Task_2_Part_3.py
@@ -1,8 +1,8 @@
 # Use the ternary operator to find the maximum of three numbers entered by the user.
 
-num1 = int(input("Enter number 1: "))
-num2 = int(input("Enter number 2: "))
-num3 = int(input("Enter number 3: "))
+num1 = float(input("Enter number 1: "))
+num2 = float(input("Enter number 2: "))
+num3 = float(input("Enter number 3: "))
 
 print("All 3 numbers are equal") if num1 == num2 == num3 \
     else print(f"Maximum of {num1}, {num2}, {num3} is:", num1 if num1 > num2 and num1 > num3
