+++ Tasks/13Oct2023/Task_2_Part_4.py
@@ -1,6 +1,6 @@
 # Develop a Python script that calculates the square and cube of a given number.
 
-num1 = int(input("Enter a number: "))
+num1 = float(input("Enter a number: "))
 print(f"Square of {num1} is:", num1 ** 2)
 print(f"Cube of {num1} is:", num1 ** 3)
